+++ src/main.js
@@ -4,7 +4,7 @@ import { createPinia } from "pinia";
 import App from "./App.vue";
 import router from "./router";
 import ElementPlus from "element-plus";
-import { i18n } from "./plugins/i18n";
+import translations from '../src/services/translations';
 
 import "./assets/styles/tailwind.css";
 import "./assets/styles/element.scss";
@@ -14,5 +14,5 @@ const app = createApp(App);
 app.use(ElementPlus);
 app.use(createPinia());
 app.use(router);
-app.use(i18n);
+app.use(translations.i18n);
 app.mount("#app");
